Multiple node unit tests
Use the etcd framework for creating a multi node cluster.

At present, these tests do not attempt to verify any of the multiple
node state.  However, this commit adds the ability too.
