+++ consensus/etcd_test.go
@@ -2,13 +2,13 @@ package consensus
 
 import (
 	"context"
-	"go.etcd.io/etcd/clientv3"
+	"github.com/stretchr/testify/assert"
+	tu "github.com/swiftstack/ProxyFS/consensus/testutils"
+	clientv3 "go.etcd.io/etcd/clientv3"
 	"go.etcd.io/etcd/clientv3/namespace"
 	"sync"
 	"testing"
 	"time"
-
-	"github.com/stretchr/testify/assert"
 )
 
 // Test basic etcd primitives
@@ -19,40 +19,33 @@ func TestEtcdAPI(t *testing.T) {
 	testTxnWatcher(t)
 	testTxnHb(t)
 	testGetMultipleKeys(t)
-
-	testCleanupTestKeys(t) // NOTE: Leave this test last!
 }
 
-// Delete test keys and recreate them using the etcd API
-func resetKeys(t *testing.T, cli *clientv3.Client, km map[string]string) {
-	assert := assert.New(t)
-	kvc := clientv3.NewKV(cli)
-	for k := range km {
-		_, _ = cli.Delete(context.TODO(), k)
-		_, err := kvc.Put(context.TODO(), k, "")
-		assert.Nil(err, "kvc.Put() returned err")
-	}
-}
+// TODO - don't start clients since want Register() to create it....
+// how create with http interface, etc?
+//
+// test:
+// 1. start 3 node cluster
+// 2. Register() with endpoints
+// 3. Do some tests with API
+// 4. Shutdown more than half of cluster
+// 6. attempt transaction and get back no leader error....
 
 // Test basic etcd API based on this link:
 // https://godoc.org/github.com/coreos/etcd/clientv3/namespace
 //
 // This test uses Put() and Get()
 func testBasicPutGet(t *testing.T) {
-	assert := assert.New(t)
 
-	// Create an etcd client - our current etcd setup does not listen on
-	// localhost.  Therefore, we pass the IP addresses used by etcd.
-	//
-	// TODO - store cli in inode volume struct.  Any issue with storing
-	// context() - comments on webpage seems to discourage it?
-	cli, err := clientv3.New(clientv3.Config{Endpoints: []string{"192.168.60.10:2379",
-		"192.168.60.11:2379", "192.168.60.12:2379"}, DialTimeout: 2 * time.Second})
-	if err != nil {
-		assert.Nil(err, "clientv3.New() returned err")
-		// handle error!
-	}
-	defer cli.Close()
+	// Start a 3 node etcd cluster
+	clus := tu.CreateCluster(t, 3)
+	defer clus.Terminate(t)
+
+	// Create a client to send/receive requests from cluster
+	cli := clus.Client(0)
+	defer tu.CleanupClient(cli, 0, clus)
+
+	assert := assert.New(t)
 
 	// Override the client interfaces using a namespace.  This
 	// allows volumes to only get relevant messages.
@@ -66,29 +59,24 @@ func testBasicPutGet(t *testing.T) {
 
 	resp, err := cli.Get(context.TODO(), "TESTabc")
 	assert.Equal("123", string(resp.Kvs[0].Value), "Get() returned different key than Put()")
-
-	// To retrieve the key from the command line you must do:
-	// # ETCDCTL_API=3 etcdctl --endpoints=http://192.168.60.11:2379 get "myVolume/abc"
+	assert.Nil(err, "Get() err != nil")
 }
 
 // Test transactions
-// TODO - change to use namespace
 func testBasicTxn(t *testing.T) {
-	assert := assert.New(t)
 
-	// Create an etcd client - our current etcd setup does not listen on
-	// localhost.  Therefore, we pass the IP addresses used by etcd.
-	cli, err := clientv3.New(clientv3.Config{Endpoints: []string{"192.168.60.10:2379",
-		"192.168.60.11:2379", "192.168.60.12:2379"}, DialTimeout: 2 * time.Second})
-	if err != nil {
-		assert.Nil(err, "clientv3.New() returned err")
-		// handle error!
-	}
-	defer cli.Close()
+	// Start a 3 node etcd cluster
+	clus := tu.CreateCluster(t, 3)
+	defer clus.Terminate(t)
+
+	// Create a client to send/receive requests from cluster
+	cli := clus.Client(0)
+	defer tu.CleanupClient(cli, 0, clus)
 
 	kvc := clientv3.NewKV(cli)
 
-	_, err = kvc.Put(context.TODO(), "TESTkey", "xyz")
+	assert := assert.New(t)
+	_, err := kvc.Put(context.TODO(), "TESTkey", "xyz")
 	if err != nil {
 		assert.Nil(err, "kvc.Put() returned err")
 	}
@@ -121,55 +109,6 @@ func testBasicTxn(t *testing.T) {
 	}
 }
 
-// TODO - must wrap with WithRequiredLeader
-func watcher(t *testing.T, cli *clientv3.Client, key string, expectedValue string,
-	swg *sync.WaitGroup, fwg *sync.WaitGroup) {
-
-	swg.Done() // The watcher is running!
-
-	assert := assert.New(t)
-	wch1 := cli.Watch(context.Background(), key)
-	for wresp1 := range wch1 {
-		for _, e := range wresp1.Events {
-			assert.Equal(expectedValue, string(e.Kv.Value),
-				"watcher saw different value than expected")
-
-			// Handle heartbeat test case
-			if string(e.Kv.Key) == "TESTHB" {
-
-				// Get current time
-				currentTime := time.Now().UTC()
-
-				// Unmarshal time HB sent
-				var sentTime time.Time
-				err := sentTime.UnmarshalText(e.Kv.Value)
-				assert.Nil(err, "UnmarshalText() failed with HB")
-
-				// Current time must be at least 500 milliseconds later
-				testTime := sentTime
-				testTime.Add(500 * time.Millisecond)
-				assert.True(currentTime.After(testTime), "testTime !> 500 milliseconds")
-			}
-		}
-
-		// The watcher has received it's event and will now return
-		fwg.Done()
-		return
-	}
-}
-
-// Start a different watcher for each key
-func startWatchers(t *testing.T, cli *clientv3.Client, km map[string]string, swg *sync.WaitGroup,
-	fwg *sync.WaitGroup) {
-	for k, v := range km {
-		swg.Add(1)
-		go watcher(t, cli, k, v, swg, fwg)
-	}
-
-	// Wait for watchers to start
-	swg.Wait()
-}
-
 // Test multiple updates of two keys in same transaction and a watcher
 // per key updated.
 //
@@ -183,28 +122,27 @@ func testTxnWatcher(t *testing.T) {
 		startWG   sync.WaitGroup // Used to sync start of watcher
 		finishWG  sync.WaitGroup // Used to block until watcher sees event
 	)
-	assert := assert.New(t)
 
-	// Create an etcd client - our current etcd setup does not listen on
-	// localhost.  Therefore, we pass the IP addresses used by etcd.
-	cli, err := clientv3.New(clientv3.Config{Endpoints: []string{"192.168.60.10:2379",
-		"192.168.60.11:2379", "192.168.60.12:2379"}, DialTimeout: 2 * time.Second})
-	if err != nil {
-		assert.Nil(err, "clientv3.New() returned err")
-		// handle error!
-	}
-	defer cli.Close()
+	// Start a 3 node etcd cluster
+	clus := tu.CreateCluster(t, 3)
+	defer clus.Terminate(t)
+
+	// Create a client to send/receive requests from cluster
+	cli := clus.Client(0)
+	defer tu.CleanupClient(cli, 0, clus)
+
+	assert := assert.New(t)
 
 	kvc := clientv3.NewKV(cli)
 
 	// Reset the test keys to be used
-	m := make(map[string]string)
-	m[testKey1] = testData1
-	m[testKey2] = testData2
-	resetKeys(t, cli, m)
+	testData := make(map[string]string)
+	testData[testKey1] = testData1
+	testData[testKey2] = testData2
+	tu.ResetKeys(t, cli, testData)
 
 	// Create watchers for each key
-	startWatchers(t, cli, m, &startWG, &finishWG)
+	tu.StartWatchers(t, cli, testData, &startWG, &finishWG)
 
 	// NOTE: "etcd does not ensure linearizability for watch operations. Users are
 	// expected to verify the revision of watch responses to ensure correct ordering."
@@ -214,9 +152,9 @@ func testTxnWatcher(t *testing.T) {
 
 	// Update multiple keys at once and guarantee get watch event for
 	// all key updates.
-	finishWG.Add(len(m))
+	finishWG.Add(len(testData))
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	_, err = kvc.Txn(ctx).
+	_, err := kvc.Txn(ctx).
 
 		// txn value comparisons are lexical
 		If(
@@ -250,31 +188,30 @@ func testTxnHb(t *testing.T) {
 		startWG  sync.WaitGroup // Used to sync start of watcher
 		finishWG sync.WaitGroup // Used to block until watcher sees event
 	)
+
+	// Start a 3 node etcd cluster
+	clus := tu.CreateCluster(t, 3)
+	defer clus.Terminate(t)
+
+	// Create a client to send/receive requests from cluster
+	cli := clus.Client(0)
+	defer tu.CleanupClient(cli, 0, clus)
+
 	assert := assert.New(t)
 
 	testHbData, errTime := time.Now().UTC().MarshalText()
 	time.Sleep(1 * time.Second)
 	assert.Nil(errTime, "time.Now.UTC().MarshalText() returned err")
 
-	// Create an etcd client - our current etcd setup does not listen on
-	// localhost.  Therefore, we pass the IP addresses used by etcd.
-	cli, err := clientv3.New(clientv3.Config{Endpoints: []string{"192.168.60.10:2379",
-		"192.168.60.11:2379", "192.168.60.12:2379"}, DialTimeout: 2 * time.Second})
-	if err != nil {
-		assert.Nil(err, "clientv3.New() returned err")
-		// handle error!
-	}
-	defer cli.Close()
-
 	kvc := clientv3.NewKV(cli)
 
 	// Reset the test keys to be used
-	m := make(map[string]string)
-	m[testHb] = string(testHbData)
-	resetKeys(t, cli, m)
+	testData := make(map[string]string)
+	testData[testHb] = string(testHbData)
+	tu.ResetKeys(t, cli, testData)
 
 	// Create watchers for each key
-	startWatchers(t, cli, m, &startWG, &finishWG)
+	tu.StartWatchers(t, cli, testData, &startWG, &finishWG)
 
 	// NOTE: "etcd does not ensure linearizability for watch operations. Users are
 	// expected to verify the revision of watch responses to ensure correct ordering."
@@ -283,9 +220,9 @@ func testTxnHb(t *testing.T) {
 	//     recognize partitioned out of cluster.
 
 	// Send HB and guarantee get watch event for HB
-	finishWG.Add(len(m))
+	finishWG.Add(len(testData))
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	_, err = kvc.Txn(ctx).
+	_, err := kvc.Txn(ctx).
 
 		// txn value comparisons are lexical
 		If(
@@ -320,33 +257,32 @@ func testGetMultipleKeys(t *testing.T) {
 		startWG   sync.WaitGroup // Used to sync start of watcher
 		finishWG  sync.WaitGroup // Used to block until watcher sees event
 	)
-	assert := assert.New(t)
 
-	// Create an etcd client - our current etcd setup does not listen on
-	// localhost.  Therefore, we pass the IP addresses used by etcd.
-	cli, err := clientv3.New(clientv3.Config{Endpoints: []string{"192.168.60.10:2379",
-		"192.168.60.11:2379", "192.168.60.12:2379"}, DialTimeout: 2 * time.Second})
-	if err != nil {
-		assert.Nil(err, "clientv3.New() returned err")
-		// handle error!
-	}
-	defer cli.Close()
+	// Start a 3 node etcd cluster
+	clus := tu.CreateCluster(t, 3)
+	defer clus.Terminate(t)
+
+	// Create a client to send/receive requests from cluster
+	cli := clus.Client(0)
+	defer tu.CleanupClient(cli, 0, clus)
+
+	assert := assert.New(t)
 
 	kvc := clientv3.NewKV(cli)
 
 	// Reset the test keys to be used
-	m := make(map[string]string)
-	m[testKey1] = string(testData1)
-	m[testKey2] = string(testData2)
-	resetKeys(t, cli, m)
+	testData := make(map[string]string)
+	testData[testKey1] = string(testData1)
+	testData[testKey2] = string(testData2)
+	tu.ResetKeys(t, cli, testData)
 
 	// Create watchers for each key
-	startWatchers(t, cli, m, &startWG, &finishWG)
+	tu.StartWatchers(t, cli, testData, &startWG, &finishWG)
 
 	// Create the keys with our data so that we can retrieve them
-	finishWG.Add(len(m))
+	finishWG.Add(len(testData))
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	_, err = kvc.Txn(ctx).
+	_, err := kvc.Txn(ctx).
 
 		// txn value comparisons are lexical
 		If(
@@ -374,21 +310,4 @@ func testGetMultipleKeys(t *testing.T) {
 	// Now get multiple keys
 	resp, err := cli.Get(context.TODO(), testStr+testKey, clientv3.WithPrefix())
 	assert.Equal(int64(2), resp.Count, "Get() returned wrong length for number of keys")
-}
-
-// Cleanup all the test keys created by these tests
-func testCleanupTestKeys(t *testing.T) {
-	assert := assert.New(t)
-
-	// Create an etcd client - our current etcd setup does not listen on
-	// localhost.  Therefore, we pass the IP addresses used by etcd.
-	cli, err := clientv3.New(clientv3.Config{Endpoints: []string{"192.168.60.10:2379",
-		"192.168.60.11:2379", "192.168.60.12:2379"}, DialTimeout: 2 * time.Second})
-	if err != nil {
-		assert.Nil(err, "clientv3.New() returned err")
-		// handle error!
-	}
-	defer cli.Close()
-
-	_, _ = cli.Delete(context.TODO(), "TEST", clientv3.WithPrefix())
 }
